Replace Double with BigDecimal (#120)
* Replace Double with BigDecimal

* rename String.toDoubleSafe

+++ app/src/main/java/dev/arkbuilders/rate/data/db/entity/RoomCurrencyRate.kt
@@ -3,11 +3,12 @@ package dev.arkbuilders.rate.data.db.entity
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import dev.arkbuilders.rate.domain.model.CurrencyCode
+import java.math.BigDecimal
 
 @Entity
 data class RoomCurrencyRate(
     @PrimaryKey
     val code: CurrencyCode,
     val currencyType: String,
-    val rate: Double,
+    val rate: BigDecimal,
 )
