+++ app/src/main/java/dev/arkbuilders/rate/data/db/entity/RoomPairAlert.kt
@@ -3,6 +3,7 @@ package dev.arkbuilders.rate.data.db.entity
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import dev.arkbuilders.rate.domain.model.CurrencyCode
+import java.math.BigDecimal
 import java.time.OffsetDateTime
 
 @Entity
@@ -11,8 +12,8 @@ data class RoomPairAlert(
     val id: Long,
     val targetCode: CurrencyCode,
     val baseCode: CurrencyCode,
-    val targetPrice: Double,
-    val startPrice: Double,
+    val targetPrice: BigDecimal,
+    val startPrice: BigDecimal,
     val alertPercent: Double?,
     val oneTimeNotRecurrent: Boolean,
     val enabled: Boolean,
